requirements.txt
asgiref==3.2.10
Django==3.1.2
django-bootstrap3==14.2.0
django-filter==2.4.0
django-tables2==2.3.2
pytz==2020.1
sqlparse==0.4.1
